Update web interface to use relays app
I had to move the elixir_ale dependency back into fw because it can only
be built for the target device. It is an optional dependency in relays,
which now defaults to it's own fake gpio implementation.

Remove all the stub modules from web interface and simply configure
relays to use the fake gpio. I had to specify that even though it's the
default because it seems the configs for the whole umbrella project are
loaded when you build one app. This is also why I didn't have to specify
the set of relay pins in web_interface... It's simply using the config
from fw